refactor: created resource-toolbar component removing code duplication and complexity. fix: link-value label formatting
