+++ apps/dsp-app/src/app/workspace/resource/properties/properties.component.scss
@@ -1,6 +1,5 @@
 @use '../../../../styles/config' as *;
 
-// toolbar
 .toolbar {
   display: flex;
   box-sizing: border-box;
@@ -10,8 +9,6 @@
   padding: 0 16px;
   width: 100%;
   color: rgba(0, 0, 0, 0.87);
-}
-.toolbar {
   background: $primary_50;
 
   &.deleted {
@@ -25,10 +22,6 @@
     white-space: nowrap;
     overflow: hidden;
     text-overflow: ellipsis;
-  }
-
-  .action button {
-    border-radius: 0;
   }
 }
