+++ apps/dsp-app/src/app/workspace/resource/properties/resource-toolbar/resource-toolbar.scss
@@ -1,0 +1,3 @@
+.action button {
+  border-radius: 0;
+}
