+++ apps/dsp-app/src/app/workspace/resource/values/link-value/link-value.component.html
@@ -1,6 +1,9 @@
 <span *ngIf="mode === 'read'; else showForm" class="read-mode-view">
   <span class="rm-value" (click)="refResClicked()" (mouseover)="refResHovered()">
-    <a class="link">{{linkLabel}}</a>
+    <a class="link"
+      ><span *ngIf="linkLabelResourceClass">{{linkLabelResourceClass}}</span
+      ><strong>{{valueFormControl.value?.label}}</strong></a
+    >
   </span>
   <span class="rm-comment" *ngIf="shouldShowComment">{{ commentFormControl.value }}</span>
 </span>
